Update patch set 1
Patch Set 1:

It's a nice code, thanks for writing that. However I believe it is a bit complex and things can be made easier. We already have a method to retrieve the profile name from site. And by design, the profile names already contain aws, libvirt, baremetal... so why don't just parse the string there to get the type?

Patch-set: 1