Update patch set 1
Patch Set 1:

I'm thinking about the case where we are reusing the same host to deploy the same site again and again (internal testing), or the case where we will want to reuse same host for several sites. Does it make sense to have this as part of fetch_requirements, or we move to another step? Or maybe add some flag , to skip the preparation of the host?

Patch-set: 1