Send build path to apply-workloads
We need to run it with sudo because previous commands
were also with sudo and there are permission problems if
not. But if we run with sudo installer tries to read
kubeconfig from /root/, not from /home/jenkins,
and it fails. So passing the build path pointing
to the right home.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>
Change-Id: I8aff76b7ece9e89258990c699b7fb36f8c70fb80