Remove bootstrap defaulting to 4.3
Change-Id: I258957cdc280e80a9a72822c89fdfe4b70bb6a17
Signed-off-by: sheetalsingala <ssingala@redhat.com>
Change-Id: Icbe34ba1f38bc7dc56e000c665bdff383334c5b7

Currently, the bootstrap machine defaults to 4.3 and the master/worker nodes conform to a version defined in 'requirements.yaml'. This is because kni-upi-lab is responsible for preparing manifests for the bootstrap node and knictl does it for master/worker nodes. This patch injects the version defined in 'requirements.yaml' into kni-upi-lab in the absence of 'releaseImageOverride'

Change-Id: I24d5124fa2b39ba2dd91065caae7204f701b706b
Signed-off-by: sheetalsingala <ssingala@redhat.com>