Remove bootstrap defaulting to 4.3
Currently, the bootstrap machine defaults to 4.3 and the master/worker nodes conform to a version defined in 'requirements.yaml'. This is because kni-upi-lab is responsible for preparing manifests for the bootstrap node and knictl does it for master/worker nodes. This patch injects the version defined in 'requirements.yaml' into kni-upi-lab in the absence of 'releaseImageOverride'

Change-Id: I0af9f56a3689436d5923f6cfb1407125403c35ab
Signed-off-by: sheetalsingala <ssingala@redhat.com>