Update patch set 3
Description "Currently, the bootstra machine defaults to 4.3 and the master/worker nodes conform to the version defined in 'requirements.yaml'. This is because kni-upi-lab is responsible for preparing manifests for the bootstrap machine and knictl does it for master/worker nodes. This patch injects the version from 'requirements.yaml' into kni-upi-lab in the absence of 'releaseImageOverride'" removed

Patch-set: 3
Patch-set-description: 
Tag: autogenerated:gerrit:setPsDescription