Run podman cleanup at start of baremetal CI job
The filesystem mounts for the containers running dnsmasq, etc,
in support of the virtual baremetal system can end up in the mounts
of the virtual machines, which means cleaning them all up needs to
be sequenced correctly: first the VMs should be removed, then the
containers.

There's an edge case where if a container is removed while the vm
is running, it will fail to remove the underlying storage, but no
longer appear in the list of containers.  This will prevent a
replacement with the same name from being created.  To handle this
case as well, we run 'podman rm --storage' after removal.  This
should allow us to clean up the host correctly even if an attempt
to clean up in the wrong order was made.

Signed-off-by: James E. Blair <jeblair@redhat.com>
Change-Id: Ie089ebab65f8a70732def5c882abf465561bebec